feat: String으로 받아 사칙연산 모두 수행

+++ src/test/java/StringCalculator/Calculator/ComplexCalculatorTest.java
@@ -9,7 +9,7 @@ public class ComplexCalculatorTest {
 
     @ParameterizedTest
     @CsvSource(value = {"1 + 1;2", "3 + 3 / 2 - 3;0", "1 + 12 - 4 / 3;3", "1 + 12 - 4 / 3 * 0;0"}, delimiter = ';')
-    public void complexInput(String input, int result) {
+    public void complexInput(String input, float result) {
         assertThat(ComplexCalculator.complexCalculate(input)).isEqualTo(result);
     }
 }
